fix(radar): the vector's range test is an int, not a widened one
clang-tidy's bugprone-implicit-widening-of-multiplication-result reads
an int64_t compared against an int product as a conversion that wanted
a wider multiplication. Nothing here needs 64 bits: the squares are
pixels on a 200 px glass.

+++ firmware/products/skyblip_go/pages/radar.cpp
@@ -443,7 +443,7 @@ void own_vector(ui::Canvas& fb, const RadarSnapshot& snap, int16_t track) {
     for (int step = 1; step <= kMinutesMarked * kStepsPerMinute; step++) {
         const HeadingUp at = on_glass_at(arc.advance(), snap, track);
         if (!inside_ring(at.right, at.ahead)) return;
-        const int64_t out = at.right * at.right + at.ahead * at.ahead;
+        const int32_t out = at.right * at.right + at.ahead * at.ahead;
         if (out < kMinuteClearPx * kMinuteClearPx) {
             pen.lift();
             continue;
